contrib(google-adk-agents): clean up draft — ava tests, lint, packaging, code fixes
Convert tests vitest→ava, add eslint override, extend tsconfig base +
references, fill package.json metadata, fix README/comment inaccuracies,
warn on streaming without streamingTopic + adaptive-heartbeat the streaming
activity, and register in CODEOWNERS + root README.
