+++ .github/CODEOWNERS
@@ -4,7 +4,8 @@
 # someone opens a pull request.
 * @temporalio/sdk
 
-/contrib/ai-sdk/           @temporalio/sdk @temporalio/ai-sdk
-/contrib/openai-agents/    @temporalio/sdk @temporalio/ai-sdk
-/contrib/strands/          @temporalio/sdk @temporalio/ai-sdk
-/contrib/workflow-streams/ @temporalio/sdk @temporalio/ai-sdk
+/contrib/ai-sdk/            @temporalio/sdk @temporalio/ai-sdk
+/contrib/google-adk-agents/ @temporalio/sdk @temporalio/ai-sdk
+/contrib/openai-agents/     @temporalio/sdk @temporalio/ai-sdk
+/contrib/strands/           @temporalio/sdk @temporalio/ai-sdk
+/contrib/workflow-streams/  @temporalio/sdk @temporalio/ai-sdk
